+++ packages/plugin-auth-wechat/src/server/wechat-auth.ts
@@ -12,6 +12,7 @@ export class WeChatAuth extends BaseAuth {
     super({
       ...config,
       userCollection: ctx.db.getCollection('users'),
+      userStatusCollection: ctx.db.getCollection('userStatuses'),
     });
   }