+++ packages/plugin-workflow-approval/src/client/user-interface/pc/block/table-initiations/CheckContent.view.tsx
@@ -8,21 +8,23 @@ import {
 import { DetailsBlockProvider } from '@tachybase/module-workflow/client';
 import { useForm } from '@tachybase/schema';
 
-import { FormBlockProvider } from '../../../../common/components/FormBlock.provider';
-import { ContextWithActionEnabled } from '../../common/WithActionEnabled.provider';
-import { useActionReminder } from '../common/hooks/useActionReminder';
-import { useActionResubmit } from '../common/hooks/useActionResubmit';
-import { useWithdrawAction } from '../common/hooks/useWithdrawAction';
-import { ProviderActionReminder } from '../common/providers/ActionReminder.provider';
-import { SchemaComponentContextProvider } from '../common/providers/SchemaComponentContextProvider';
-import { WithdrawActionProvider } from '../common/providers/WithdrawAction.provider';
+import {
+  ActionBarProvider,
+  ContextWithActionEnabled,
+  FormBlockProvider,
+  ProviderActionReminder,
+  ProviderActionResubmit,
+  SchemaComponentContextProvider,
+  useActionReminder,
+  useActionResubmit,
+  useDestroyAction,
+  useFormBlockProps,
+  useSubmitUpdate,
+  useWithdrawAction,
+  WithdrawActionProvider,
+} from '../../../../common';
+import { ProviderApplyActionStatus } from '../../../../common/contexts/ApplyActionStatus.provider';
 import { getSchemaCheckContent } from './CheckContent.schema';
-import { useDestroyAction } from './hooks/useDestroyAction';
-import { useFormBlockProps } from './hooks/useFormBlockProps';
-import { useSubmitUpdate } from './hooks/useSubmitUpdate';
-import { ActionBarProvider } from './providers/ActionBar.provider';
-import { ProviderActionResubmit } from './providers/ActionResubmit.provider';
-import { ProviderApplyActionStatus } from './providers/ApplyActionStatus.provider';
 
 export const ViewCheckContent = (props) => {
   const { approval, workflow } = props;